requirements.txt
streamlit==1.32.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.7
Pillow==10.0.0
numpy==1.24.0
plotly==5.17.0
gdown==4.7.3